Combine the call and pads screen
